requirements/base.txt
# pytz==2021.1  # https://github.com/stub42/pytz
# python-slugify==5.0.2  # https://github.com/un33k/python-slugify
# Pillow==8.2.0  # https://github.com/python-pillow/Pillow
rcssmin==1.0.6  # https://github.com/ndparker/rcssmin
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.2.0  # https://github.com/evansd/whitenoise
redis==3.5.3  # https://github.com/antirez/redis
range-key-dict==1.1.0

# Django
# ------------------------------------------------------------------------------
django==3.2.3  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
# django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.44.0  # https://github.com/pennersr/django-allauth
django-foundation-formtags==0.1.0
django-compressor==2.4.1  # https://github.com/django-compressor/django-compressor
django-libsass==0.8
django-redis==4.12.1  # https://github.com/niwinz/django-redis
django-distill==2.6
django-crispy-forms==1.11.2
django-geojson==3.1.0

# Django REST Framework
djangorestframework==3.12.4  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
djangorestframework-mvt==0.2.5
